+++ Panther/include/Diagnostic.h
@@ -595,32 +595,19 @@ namespace pcit::panther{
 		struct Info{
 			std::string message;
 			Location location;
-			evo::SmallVector<Info> sub_infos;
+			evo::SmallVector<Info> subInfos;
 
-			Info(std::string&& _message) : message(std::move(_message)), location(), sub_infos() {
+			Info(std::string&& _message) : message(std::move(_message)), location(), subInfos() {
 				// _debug_analyze_message(this->message);
 			}
 			Info(std::string&& _message, Location loc) 
-				: message(std::move(_message)), location(loc), sub_infos() {
+				: message(std::move(_message)), location(loc), subInfos() {
 				// _debug_analyze_message(this->message);
 			}
-			Info(std::string&& _message, Location loc, evo::SmallVector<Info>&& _sub_infos)
-				: message(std::move(_message)), location(loc), sub_infos(std::move(_sub_infos)) {
+			Info(std::string&& _message, Location loc, evo::SmallVector<Info>&& _subInfos)
+				: message(std::move(_message)), location(loc), subInfos(std::move(_subInfos)) {
 				// _debug_analyze_message(this->message);
 			}
-
-			// Info(const Info& rhs) : message(rhs.message), location(rhs.location), sub_infos(rhs.sub_infos) {}
-
-			// auto operator=(const Info& rhs) -> Info& {
-			// 	this->message = rhs.message;
-			// 	this->location = rhs.location;
-			// 	this->sub_infos = rhs.sub_infos;
-				
-			// 	return *this;
-			// }
-
-			// Info(Info&& rhs)
-			// 	: message(std::move(rhs.message)), location(rhs.location), sub_infos(std::move(rhs.sub_infos)) {}
 		};
 
 		Level level;